refactor: move modal components files
