+++ src/actions/modals/components/test/group-calendars-list-item.test.tsx
@@ -7,8 +7,8 @@ import React from 'react';
 
 import { ZIMBRA_STANDARD_COLORS } from '@zextras/carbonio-ui-commons';
 
-import { TEST_SELECTORS } from '../../../constants/test-utils';
-import { generateGroupCalendar } from '../../../test/generators/group';
+import { TEST_SELECTORS } from '../../../../constants/test-utils';
+import { generateGroupCalendar } from '../../../../test/generators/group';
 import { GroupCalendarsListItem } from '../group-calendars-list-item';
 import { setupTest, screen } from '@test-setup';
